added check for possible pacman/kalu conflict
Preferences, Upgrades: Check for pacman/kalu conflict (CheckPacmanConflict)
When enabled, kalu will check if there's an upgrade of pacman to be done, and if so if it is likely to prevent the sysupgrade due to kalu's dependency (i.e. if pacman's major or minor version number goes up, meaning probable changes in libalpm's API).
If so, an extra button will be featured on the notification, to show a message about this issue and how to perform the system upgrade.