When buttons for Upgrades & AUR were both used/clicked at the same time, kalu would run a check after each, fixed; Closes #8
Usually when kalu is busy nothing else to get it busy can be started. One exception is starting Upgrades & AUR from notification buttons. In such a case, when the first one is done, kalu would run a check (to refresh) then do it again after the second one.
And using kalu's updater its state could go back to not busy even though it or the other process is still running, possibly leading to more trouble.
kalu will now wait for both to be done to go back to be non-busy & do whatever needs to be done.