Fix re-show notifications which would ignore conflicting files error
When the list of packages to upgrade couldn't be compiled due to file
conflict, a "special" notif is shown, with an error message but also
the "Upgrade system" button.
This notification wasn't registered in the last_notifs list, and
re-show would ignore it (e.g. say "no notifications to show")