Set icon & tooltip properly when conflict makes upgrades number unknown
When the number of upgrades isn't known due to a conflict, kalu's icon
would not be set blue (i.e. upgrades available) and remain as it was
after last conflict-free check, the tooltip would also read whatever
was set from the last ran check without conflict (if any).

This could result in the previous number of updates (before the conflict
was introduced), or just none - making it look as if no upgrades are
available, which is just wrong.

Now the icon is always blue, and the tooltip say: "upgrades available
(unknown number due to conflict)"