configure.ac: Better handling of --enable-git-version
Multiple messages were put on the same line, so the output could be a
bit confusing. Also, it didn't fail on error, i.e. if git or .git were
missing the configure still ran successfully, but git version wasn't
enabled.

Now all messages are on their own line, if git or .git is not found,
an error occurs, and on success the version ("kalu version x.y.z") is
suffixed with "-git"