Fix empty notifications when '&' was used in text
Seems that because it uses a subset of HTML, the text in notifications should
not use '&' directly, but encode it as &amp; (else, the text of notification
is just empty, at least with notification-daemon and xfce4-notifyd)

So, when creating new notifications, we make sure to "convert" those '&' since
there can be some in e.g. a package description, which could be used as part
of the template (text is untouched on CLI output).