Refactor escaping for notifications (Fix #3)
We now escape a few more characters, thus fixing a bug where notification-daemon
would not show the summary text if a '<' was used.
Escaping is only done on the fields where it applies (i.e. not for sizes, etc)
when send to notification (i.e. not for tty output), thus assuming that
user-provided templates are correctly escaped (to allow the use of tags, for
bold, etc).