Support ALPM's new events PKGDOWNLOAD_*
This allows to know when the download of a package file begins, and when
it ends (whether successully or not).

So we can move some code out of every progress update, to the start of
a download, or the end of it.

This will also fix a bug where if a download failed, its progress bar
would never go away (since it was based on reaching 100%). Now not only
does it go away, but the size will show up in red afterwards, to
indicate there was a failure.