Add notif for packages not found in AUR
When checking the AUR, any and all packages not found will now be listed
in an additionnal notification. This allows to catch e.g. when a package
might have been renamed/removed.

Note that such packages should be put on the ignore list, to avoid useless
querying of the AUR.