Add KDE's StatusNotifierItem support
With Plasma Next KDE doesn't support the good old xembed systray, and so
GtkStatusIcon are "useless."

Adding a new configure option --enable-status-notifier to enable support
of KDE's interface via our small library, statusnotifier.

When used, kalu will try to register a StatusNotifierItem first. If it
fails, it will fallback to using the systray.