prefs: Unchecking a template field wasn't always remembered
When unchecking a template field (to have it use its fallback value) it
wouldn't be remembered when said field has a hard-coded default value
in kalu, which would then be restored. (E.g. the Title field for packages
not found in the AUR)

Now such fields are saved to the .conf file without a value (e.g. "Title=")
to preserve their disabled/falling-back state.

Fixes #17