updater: Add ability to abort a transaction
kalu-dbus now handles SIGINT to abort the current ALPM transaction. Said
transaction (i.e. the sysupgrade, or download-only) will now be performed
in a new thread, so that kalu-dbus can still process other method calls
at the same time, specifically the new method Abort.

Abort() will send SIGINT into the sysupgrade thread, so that the signal
may be caught be ALPM (in case of a download, to properly abort it) before
it is raised again, and then we'll interrupt the transaction.

In effect, this is like pressing ^C from your pacman-running terminal.