Keep tmp dbs around to speed things up/save bandwidth
Instead of removing the tmp dbs after the checks/simulation, we now keep
them around. That way, or the next checks/simulation we can re-use them
and might alreayd be up-to-date (even though the system's ones aren't).

If the DB path changed or the folder is gone we just re-create tmp dbs,
and of course we clean things up on app end.