watched: Add support for repo/package
This allows to have kalu watched for a given package in a specific repo, instead
of using the usual "first repo where package is found" logic.

This could be useful when e.g. one has a repo like testing enabled, but wants
to watch packages from core.

Note that the package name (i.e. $PKG) will be kept as "repo/package" to
indicate the "restriction" that was in effect.