updater: Fix color glitch when a download fails & resumes
Since 55468d2d and the support for multiple downloads per packages (via
deltas) there was an issue when a download fails, then resumes (e.g. on
another mirror), as the total size downloaded wasn't matching the expected
one, hence the download not considered (fully) done.

This was because even on download_failed the size of the download was
added as done, which obviously isn't right. In order to make it work as
expected, we need to then only add the actually downloaded size, but more
importantly we need to handle resuming from another mirror, which we can only do
from on_download() as there's no event generated in such a case.

Thanks to jghodd; Closes #45.