news: Fix issue if title needs trimming..
..that is, if a news title (used to determine whether it has been read
or not) could be trimmedg, there was an issue because titles read from
news.conf are trimmed, so the two wouldn't match, and the news
incorrectly considered unread.

To fix this we trim the titles read from the XML (RSS feed) as well.

Thanks to Ada Joule; Also adam777 for the report.

Fixes #57